requirements.txt
# ── Core Dashboard ────────────────────────────────────────────────
streamlit>=1.32.0

# ── Data Processing ───────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0

# ── Visualization ─────────────────────────────────────────────────
plotly>=5.18.0

# ── AI Insights ───────────────────────────────────────────────────
openai>=1.12.0

# ── Meta API ──────────────────────────────────────────────────────
facebook-business>=19.0.0
requests>=2.31.0

# ── Environment ───────────────────────────────────────────────────
python-dotenv>=1.0.0
Pillow==10.2.0

gspread>=6.0.0
google-auth>=2.0.0